# change to single seed of RNG in init()