Having all the extraneous calls to Seed() causes almost a 5x performance loss in benchmarks on my 2011 iMac ( 4 cores of Core i5)
rand.Seed() should be called once (often placed in init())

Original version:
 go test -bench .
testing: warning: no tests to run
PASS
BenchmarkGetPi	       1	12959684666 ns/op
BenchmarkGetPiMulti	       1	4497506505 ns/op
testing: BenchmarkGetPiMulti left GOMAXPROCS set to 4
ok  	github.com/billhathaway/go-parallelism-monte-carlo-demo	17.467s


new version:
go test -bench .
testing: warning: no tests to run
PASS
BenchmarkGetPi	      20	  68930991 ns/op
BenchmarkGetPiMulti	       3	 354690543 ns/op
testing: BenchmarkGetPiMulti left GOMAXPROCS set to 4
ok  	github.com/billhathaway/go-parallelism-monte-carlo-demo	3.587s