@soroushjp 2015-02-07
You're absolutely right. I get more than a 150x speedup from these changes, merging now. Will rerun benchmarks and update blog post shortly also. Thank you Bill.